WS1: extract model-map storage helpers into map-storage.ts
Second slice of the modelmap.tsx split (1,093 -> 1,060). Move the
sessionStorage persistence + cache-key helpers (getMapStorageKey,
getStoredHeatmapMode, getStoredCurrentTime, getPeopleMapCacheKey) into
src/features/model-map/map-storage.ts and import them back. HEATMAP_MODES is
now referenced only inside that module, so it drops out of the component's
map-constants import.

Behavior-preserving: tsc --noEmit clean, pnpm test 23/23.
